[reorg] Move more GPOS lookups to new directory.
https://github.com/harfbuzz/harfbuzz/commit/197d9a5c9
