+++ src/hb/ot/layout/GPOS/mod.rs
@@ -1,4 +1,9 @@
+pub(crate) mod anchor;
 pub(crate) mod cursive_pos;
+pub(crate) mod mark_array;
+pub(crate) mod mark_base_pos;
+pub(crate) mod mark_lig_pos;
+pub(crate) mod mark_mark_pos;
 pub(crate) mod pair_pos;
 pub(crate) mod pos_lookup;
 pub(crate) mod single_pos;
