+++ src/hb/ot_layout_gpos_table.rs
@@ -1,12 +1,12 @@
-use ttf_parser::gpos::*;
-use ttf_parser::opentype_layout::LookupIndex;
 use super::buffer::*;
 use super::hb_font_t;
 use super::ot_layout::*;
-use super::ot_layout_common::{lookup_flags, PositioningLookup, PositioningTable};
-use super::ot_layout_gsubgpos::{Apply, OT::hb_ot_apply_context_t, skipping_iterator_t};
+use super::ot_layout_common::{PositioningLookup, PositioningTable};
+use super::ot_layout_gsubgpos::{Apply, OT::hb_ot_apply_context_t};
 use super::ot_shape_plan::hb_ot_shape_plan_t;
 use crate::Direction;
+use ttf_parser::gpos::*;
+use ttf_parser::opentype_layout::LookupIndex;
 
 pub fn position(plan: &hb_ot_shape_plan_t, face: &hb_font_t, buffer: &mut hb_buffer_t) {
     apply_layout_table(plan, face, buffer, face.gpos.as_ref());
@@ -143,220 +143,13 @@ impl<'a> LayoutTable for PositioningTable<'a> {
     }
 }
 
-impl Apply for MarkToBaseAdjustment<'_> {
-    fn apply(&self, ctx: &mut hb_ot_apply_context_t) -> Option<()> {
-        let buffer = &ctx.buffer;
-        let mark_glyph = ctx.buffer.cur(0).as_glyph();
-        let mark_index = self.mark_coverage.get(mark_glyph)?;
-
-        // Now we search backwards for a non-mark glyph
-        let mut iter = skipping_iterator_t::new(ctx, buffer.idx, 1, false);
-        iter.set_lookup_props(u32::from(lookup_flags::IGNORE_MARKS));
-
-        let info = &buffer.info;
-        loop {
-            let mut unsafe_from = 0;
-            if !iter.prev(Some(&mut unsafe_from)) {
-                ctx.buffer
-                    .unsafe_to_concat_from_outbuffer(Some(unsafe_from), Some(ctx.buffer.idx + 1));
-                return None;
-            }
-
-            // We only want to attach to the first of a MultipleSubst sequence.
-            // https://github.com/harfbuzz/harfbuzz/issues/740
-            // Reject others...
-            // ...but stop if we find a mark in the MultipleSubst sequence:
-            // https://github.com/harfbuzz/harfbuzz/issues/1020
-            let idx = iter.index();
-            if !_hb_glyph_info_multiplied(&info[idx])
-                || _hb_glyph_info_get_lig_comp(&info[idx]) == 0
-                || idx == 0
-                || _hb_glyph_info_is_mark(&info[idx - 1])
-                || _hb_glyph_info_get_lig_id(&info[idx])
-                    != _hb_glyph_info_get_lig_id(&info[idx - 1])
-                || _hb_glyph_info_get_lig_comp(&info[idx])
-                    != _hb_glyph_info_get_lig_comp(&info[idx - 1]) + 1
-            {
-                break;
-            }
-            iter.reject();
-        }
-
-        // Checking that matched glyph is actually a base glyph by GDEF is too strong; disabled
-
-        let iter_idx = iter.index();
-        let base_glyph = info[iter_idx].as_glyph();
-        let Some(base_index) = self.base_coverage.get(base_glyph) else {
-            ctx.buffer
-                .unsafe_to_concat_from_outbuffer(Some(iter_idx), Some(buffer.idx + 1));
-            return None;
-        };
-
-        self.marks
-            .apply(ctx, self.anchors, mark_index, base_index, iter_idx)
-    }
-}
-
-impl Apply for MarkToLigatureAdjustment<'_> {
-    fn apply(&self, ctx: &mut hb_ot_apply_context_t) -> Option<()> {
-        let buffer = &ctx.buffer;
-        let mark_glyph = ctx.buffer.cur(0).as_glyph();
-        let mark_index = self.mark_coverage.get(mark_glyph)?;
-
-        // Now we search backwards for a non-mark glyph
-        let mut iter = skipping_iterator_t::new(ctx, buffer.idx, 1, false);
-        iter.set_lookup_props(u32::from(lookup_flags::IGNORE_MARKS));
-
-        let mut unsafe_from = 0;
-        if !iter.prev(Some(&mut unsafe_from)) {
-            ctx.buffer
-                .unsafe_to_concat_from_outbuffer(Some(unsafe_from), Some(ctx.buffer.idx + 1));
-            return None;
-        }
-
-        // Checking that matched glyph is actually a ligature by GDEF is too strong; disabled
-
-        let iter_idx = iter.index();
-        let lig_glyph = buffer.info[iter_idx].as_glyph();
-        let Some(lig_index) = self.ligature_coverage.get(lig_glyph) else {
-            ctx.buffer
-                .unsafe_to_concat_from_outbuffer(Some(iter_idx), Some(buffer.idx + 1));
-            return None;
-        };
-        let lig_attach = self.ligature_array.get(lig_index)?;
-
-        // Find component to attach to
-        let comp_count = lig_attach.rows;
-        if comp_count == 0 {
-            ctx.buffer
-                .unsafe_to_concat_from_outbuffer(Some(iter_idx), Some(buffer.idx + 1));
-            return None;
-        }
-
-        // We must now check whether the ligature ID of the current mark glyph
-        // is identical to the ligature ID of the found ligature.  If yes, we
-        // can directly use the component index.  If not, we attach the mark
-        // glyph to the last component of the ligature.
-        let lig_id = _hb_glyph_info_get_lig_id(&buffer.info[iter_idx]);
-        let mark_id = _hb_glyph_info_get_lig_id(&buffer.cur(0));
-        let mark_comp = u16::from(_hb_glyph_info_get_lig_comp(buffer.cur(0)));
-        let matches = lig_id != 0 && lig_id == mark_id && mark_comp > 0;
-        let comp_index = if matches {
-            mark_comp.min(comp_count)
-        } else {
-            comp_count
-        } - 1;
-
-        self.marks
-            .apply(ctx, lig_attach, mark_index, comp_index, iter_idx)
-    }
-}
-
-impl Apply for MarkToMarkAdjustment<'_> {
-    fn apply(&self, ctx: &mut hb_ot_apply_context_t) -> Option<()> {
-        let buffer = &ctx.buffer;
-        let mark1_glyph = ctx.buffer.cur(0).as_glyph();
-        let mark1_index = self.mark1_coverage.get(mark1_glyph)?;
-
-        // Now we search backwards for a suitable mark glyph until a non-mark glyph
-        let mut iter = skipping_iterator_t::new(ctx, buffer.idx, 1, false);
-        iter.set_lookup_props(ctx.lookup_props & !u32::from(lookup_flags::IGNORE_FLAGS));
-
-        let mut unsafe_from = 0;
-        if !iter.prev(Some(&mut unsafe_from)) {
-            ctx.buffer
-                .unsafe_to_concat_from_outbuffer(Some(unsafe_from), Some(ctx.buffer.idx + 1));
-            return None;
-        }
-
-        let iter_idx = iter.index();
-        if !_hb_glyph_info_is_mark(&buffer.info[iter_idx]) {
-            ctx.buffer
-                .unsafe_to_concat_from_outbuffer(Some(iter_idx), Some(buffer.idx + 1));
-            return None;
-        }
-
-        let id1 = _hb_glyph_info_get_lig_id(buffer.cur(0));
-        let id2 = _hb_glyph_info_get_lig_id(&buffer.info[iter_idx]);
-        let comp1 = _hb_glyph_info_get_lig_comp(buffer.cur(0));
-        let comp2 = _hb_glyph_info_get_lig_comp(&buffer.info[iter_idx]);
-
-        let matches = if id1 == id2 {
-            // Marks belonging to the same base
-            // or marks belonging to the same ligature component.
-            id1 == 0 || comp1 == comp2
-        } else {
-            // If ligature ids don't match, it may be the case that one of the marks
-            // itself is a ligature.  In which case match.
-            (id1 > 0 && comp1 == 0) || (id2 > 0 && comp2 == 0)
-        };
-
-        if !matches {
-            ctx.buffer
-                .unsafe_to_concat_from_outbuffer(Some(iter_idx), Some(buffer.idx + 1));
-            return None;
-        }
-
-        let mark2_glyph = buffer.info[iter_idx].as_glyph();
-        let mark2_index = self.mark2_coverage.get(mark2_glyph)?;
-
-        self.marks
-            .apply(ctx, self.mark2_matrix, mark1_index, mark2_index, iter_idx)
-    }
-}
-
-trait MarkArrayExt {
-    fn apply(
-        &self,
-        ctx: &mut hb_ot_apply_context_t,
-        anchors: AnchorMatrix,
-        mark_index: u16,
-        glyph_index: u16,
-        glyph_pos: usize,
-    ) -> Option<()>;
-}
-
-impl MarkArrayExt for MarkArray<'_> {
-    fn apply(
-        &self,
-        ctx: &mut hb_ot_apply_context_t,
-        anchors: AnchorMatrix,
-        mark_index: u16,
-        glyph_index: u16,
-        glyph_pos: usize,
-    ) -> Option<()> {
-        // If this subtable doesn't have an anchor for this base and this class
-        // return `None` such that the subsequent subtables have a chance at it.
-        let (mark_class, mark_anchor) = self.get(mark_index)?;
-        let base_anchor = anchors.get(glyph_index, mark_class)?;
-
-        let (mark_x, mark_y) = mark_anchor.get(ctx.face);
-        let (base_x, base_y) = base_anchor.get(ctx.face);
-
-        ctx.buffer
-            .unsafe_to_break(Some(glyph_pos), Some(ctx.buffer.idx + 1));
-
-        let idx = ctx.buffer.idx;
-        let pos = ctx.buffer.cur_pos_mut();
-        pos.x_offset = base_x - mark_x;
-        pos.y_offset = base_y - mark_y;
-        pos.set_attach_type(attach_type::MARK);
-        pos.set_attach_chain((glyph_pos as isize - idx as isize) as i16);
-
-        ctx.buffer.scratch_flags |= HB_BUFFER_SCRATCH_FLAG_HAS_GPOS_ATTACHMENT;
-        ctx.buffer.idx += 1;
-
-        Some(())
-    }
-}
-
 pub mod attach_type {
     pub const MARK: u8 = 1;
     pub const CURSIVE: u8 = 2;
 }
 
 /// Just like TryFrom<N>, but for numeric types not supported by the Rust's std.
-pub trait TryNumFrom<T>: Sized {
+pub(crate) trait TryNumFrom<T>: Sized {
     /// Casts between numeric types.
     fn try_num_from(_: T) -> Option<Self>;
 }
@@ -383,7 +176,7 @@ impl TryNumFrom<f32> for i32 {
     }
 }
 
-trait DeviceExt {
+pub(crate) trait DeviceExt {
     fn get_x_delta(&self, face: &hb_font_t) -> Option<i32>;
     fn get_y_delta(&self, face: &hb_font_t) -> Option<i32>;
 }
